Fix a bug affecting finalizers
Previously we were tracking an incorrect address (the address of the ctypes pointer object, rather than the address of the C struct it points to). Usually this still works because it's unlikely that `addressof(ptr)` will be reassigned after `ptr` gets garbage collected and before the finalizer gets called.

I've also added another logging statement when the finalizer is created (which should be immediately after the pointer is allocated).

PiperOrigin-RevId: 181799057