Randomize the cheetah's internal joints at the start of each episode
This randomization has quite a subtle effect on the initial state since we run the simulation forward for 2 seconds at the start of each episode, allowing the joints to settle in a similar position across episodes.

I also added a regression test that checks that two consecutive initial states have different observations.

Closes #86

PiperOrigin-RevId: 243584776