Fix a bug in `egl_ext.py` due to mismatching `EGLDeviceExt` pointer types
As of PyOpenGL v3.1.4, `OpenGL.EGL` has its own `EGLDevicesEXT` member, which overrides our assignment of `EGLDeviceEXT = ctypes.c_void_p` when we later do `from OpenGL.EGL import *`. Consequently we end up with mismatching pointer types in the `CFUNCTYPE` wrapper for `_eglQueryDevicesEXT` and in the body of `eglQueryDevicesEXT`. I have bumped the required version of PyOpenGL to 3.4.1 and switched to using `EGL.EGLDevicesEXT` instead of `ctypes.c_void_p`.

Closes #123

PiperOrigin-RevId: 288398964
Change-Id: If5b1a9dc929d256652e0b50c60adb43d1f98b002