Allow legacy_base.Walker to use multiple initializers.
Currently, Walker only accepts a single initializer. This means that modular initialization components (such as setting a walker-specific random pose followed by a random rotation) must either be redundantly implemented in each walker-specific initializer, or must be combined with a wrapping initializer.

This change allows Walker to optionally accept multiple initializers, which are then executed sequentially to initializer the walker.

PiperOrigin-RevId: 299330267
Change-Id: I84d0cae2fca4972458254d87306cd165b197d6ff