Improve lighting in soccer pitch.
Plane geoms in mujoco don't actually have a height, they extend infinitely downward. Instead, the third number in the size argument represents the size of the grid which is used for rendering lighting. Lighting values for the plane are calculated at each grid vertex and interpolated in between. The current large grid size is what's causing the lighting artifacts we've been seeing in soccer videos.

This CL reduces the grid size, which makes lighting on the ground look much more realistic. It also renames the grid size variable to reflect its true purpose.]

PiperOrigin-RevId: 300565743
Change-Id: I9c5810cbaec576eeb012f584ad7c1dc659faa4cd