Use +/-`mjMAXVAL` instead of `inf` in bounds for unlimited actuators in `action_spec`
`mjMAXVAL` is a more correct bound, since setting `ctrl` to values that are greater than this would result in
```python
PhysicsError: Physics state is invalid. Warning(s) raised: mjWARN_BADCTRL
```
when the environment is stepped. Unbounded actions are allowed in general, provided that they are not used to set values in `ctrl` directly.
PiperOrigin-RevId: 339746781
Change-Id: Ie108dcf66df298652f635a421ed6d9a662101d47