Use built-in types instead of numpy aliases
Numpy types like `np.{bool,int,float}` are just type aliases of the
built-in types. The built-in types should be preferred, as discussed
e.g. in the [numpy 1.20.0 release notes](https://numpy.org/devdocs/release/1.20.0-notes.html#using-the-aliases-of-builtin-types-like-np-int-is-deprecated)

This commit replaces all the numpy type alises with their corresponding
built-in type.