Revert commit 73c31ac21d1214f0a2eadfc8540a7f76796955cc.
That commit broke rendering on macOS. `glfwInitHint` only sets internal hints in GLFW and a subsequent call to `glfwInit` is still required. However, `glfwInitHint(GLFW_COCOA_MENUBAR, 0); glfwInit();` still results in a dock icon being created, so we are reverting that commit entirely until we figure out a way to actually disable it.

Closes #239

PiperOrigin-RevId: 425341097
Change-Id: If87643db372d4ec0a5237deb6c0b3effecdd7dc0