Remove some of the code generation from dm_control.
After the migration to the new MuJoCo python bindings, there is not need to use the `types` or `wrappers` libraries from `mjbindings`, and using them is a likely mistake.
Other mjbindings libraries remain, to support backwards compatible use. `functions` exports the mjlib symbol, which is an alias to mujoco.
enums and constants will continue to work as before.

PiperOrigin-RevId: 433461696
Change-Id: I99de9731d91ebb39c35002b0144e5f57dc12f4a9