Speed up warning checking in dm_control.
`MjData.warning.number` is a property, and accessing it incurs a roundtrip penalty to pybind11. Instead, we now grab a reference to the returned array once. Also revert to using `np.copyto` to copy warning states -- we used to do this before the migration to the new bindings, and profiling shows that it is indeed faster.

PiperOrigin-RevId: 465284823
Change-Id: Idf0b2c26134db9fadd52d52c313ae3363c0da818