Make sure that `contact_force` always returns the correct values.
The `physics.step` function in `dm_control` runs `mj_step2` followed by `mj_step1`. This leaves the physics in a state where `mj_fwdPosition` has been called, and therefore the contact frames have been updated, but`mj_fwdConstraint` has not, and therefore the contact forces are stale. This change forces the portion of `mj_step2` that is needed to guarantee correct contact forces to be always be run as part of the `contact_force` function.

Fixes https://github.com/deepmind/mujoco/issues/13.

PiperOrigin-RevId: 490922782
Change-Id: I0a3998ff4759aee70321a636c189e44de2f3eca5