Allow full model-relative paths to be used as keys in asset dicts.
Allowing keys in this form means the asset dict can be constructed without needing to first parse the XML to determine whether any special directories have been specified in the compiler options.

For example, given an asset in `assets/textures/texture.png` and an `assetdir` set to "assets", this asset previously had to be keyed in the assets dict by "textures/texture.png". After this CL, it's also permitted to use "assets/textures/texture.png".

PiperOrigin-RevId: 807216952
Change-Id: Ie1b8680eec447562675e4ca83aa15b1c4511461e